Allow specifying `rescue_from` not accepting cause
`ActiveSupport::Rescuable.rescue_from` will select handlers by
first finding handlers for the specific exception. If no handler
is found and the exception has a `cause`, it will select handlers
for the cause, and so on until a handler is found or the
exception cause chain is exhausted.

Often, the exception cause has nothing to do with the original
exception. This behaviour is then unexpected. This can cause
some problems and debug headaches when unrelated errors that
get caught in it.

```ruby
rescue_from ActiveResource::ClientError do |error|
  if error.response.code == "402"
    render :payment_required
  else
    render :error
  end
end

def fetch_remote_resource
  main_resource = RemoteResource.find
  # load more resources
rescue ActiveResource::ResourceNotFound => e
  log("failed to load associations of #{main_resource.id}")
end
```

In the previous example, if the load of `main_resource` fails with
a 404 (due to a bug), the `rescue_from` will be called with a
`NoMethodError`. `NoMethodError` obviously doesn't respond to
`response.code`. The handler block will then raise with a new
`NoMethodError`.

I understand the rationale of matching `exception.cause` in cases
where the handler is very generic, like logging or rendering an
error message. For specific handlers, it means the handler needs
to check if the handler is exception is actually what's expected.

In the previous example, I think it would have been sane to let
the first `NoMethodError` bubble up to the middlewares and get
caught by `ActionDispatch::DebugExceptions` or equivalent.